+++ config/configtest/configtest_test.go
@@ -16,70 +16,12 @@ package configtest
 
 import (
 	"io"
-	"path"
 	"strings"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"go.opentelemetry.io/collector/component/componenttest"
-	"go.opentelemetry.io/collector/config"
 )
-
-func TestLoadConfig(t *testing.T) {
-	factories, err := componenttest.NopFactories()
-	assert.NoError(t, err)
-
-	cfg, err := LoadConfig(path.Join("testdata", "config.yaml"), factories)
-	require.NoError(t, err)
-
-	// Verify extensions.
-	require.Len(t, cfg.Extensions, 2)
-	assert.Contains(t, cfg.Extensions, config.NewComponentID("nop"))
-	assert.Contains(t, cfg.Extensions, config.NewComponentIDWithName("nop", "myextension"))
-
-	// Verify receivers
-	require.Len(t, cfg.Receivers, 2)
-	assert.Contains(t, cfg.Receivers, config.NewComponentID("nop"))
-	assert.Contains(t, cfg.Receivers, config.NewComponentIDWithName("nop", "myreceiver"))
-
-	// Verify exporters
-	assert.Len(t, cfg.Exporters, 2)
-	assert.Contains(t, cfg.Exporters, config.NewComponentID("nop"))
-	assert.Contains(t, cfg.Exporters, config.NewComponentIDWithName("nop", "myexporter"))
-
-	// Verify Processors
-	assert.Len(t, cfg.Processors, 2)
-	assert.Contains(t, cfg.Processors, config.NewComponentID("nop"))
-	assert.Contains(t, cfg.Processors, config.NewComponentIDWithName("nop", "myprocessor"))
-
-	// Verify service.
-	require.Len(t, cfg.Service.Extensions, 1)
-	assert.Contains(t, cfg.Service.Extensions, config.NewComponentID("nop"))
-	require.Len(t, cfg.Service.Pipelines, 1)
-	assert.Equal(t,
-		&config.Pipeline{
-			Receivers:  []config.ComponentID{config.NewComponentID("nop")},
-			Processors: []config.ComponentID{config.NewComponentID("nop")},
-			Exporters:  []config.ComponentID{config.NewComponentID("nop")},
-		},
-		cfg.Service.Pipelines[config.NewComponentID("traces")],
-		"Did not load pipeline config correctly")
-}
-
-func TestLoadConfigAndValidate(t *testing.T) {
-	factories, err := componenttest.NopFactories()
-	assert.NoError(t, err)
-
-	cfgValidate, errValidate := LoadConfigAndValidate(path.Join("testdata", "config.yaml"), factories)
-	require.NoError(t, errValidate)
-
-	cfg, errLoad := LoadConfig(path.Join("testdata", "config.yaml"), factories)
-	require.NoError(t, errLoad)
-
-	assert.Equal(t, cfg, cfgValidate)
-}
 
 func TestCheckConfigStructPointerAndValue(t *testing.T) {
 	config := struct {